+++ packages/wonder-blocks-tooltip/src/components/tooltip.js
@@ -122,6 +122,30 @@ type DefaultProps = {|
     placement: $PropertyType<Props, "placement">,
 |};
 
+/**
+ * Use a tooltip to help describe an on screen object.
+ *
+ * Tooltips:
+ *  - contain text
+ *  - (optional) contain small graphic elements to complement the text
+ *  - appear on hover or focus (for non-assistive tech keyboard users)
+ *  - must have a tail that points to a parent object
+ *  - DO NOT include actions
+ *
+ * For more rich content see Popovers, for taking action on an object, see
+ * Snackbars (proposed).
+ *
+ * ### Usage
+ *
+ * ```jsx
+ * import Tooltip from "@khanacademy/wonder-blocks-tooltip";
+ *
+ * <Tooltip content="This is a text tooltip">
+ *  Tooltip anchor
+ * </Tooltip>
+ * ```
+ *
+ */
 export default class Tooltip extends React.Component<Props, State> {
     static defaultProps: DefaultProps = {
         forceAnchorFocusivity: true,