+++ packages/wonder-blocks-birthday-picker/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @khanacademy/wonder-blocks-birthday-picker
 
+## 1.2.21
+
+### Patch Changes
+
+-   @khanacademy/wonder-blocks-dropdown@2.9.4
+
+## 1.2.20
+
+### Patch Changes
+
+-   @khanacademy/wonder-blocks-dropdown@2.9.3
+
 ## 1.2.19
 
 ### Patch Changes