+++ packages/wonder-blocks-popover/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @khanacademy/wonder-blocks-popover
 
+## 1.1.22
+
+### Patch Changes
+
+-   Updated dependencies [13cdc7fe]
+    -   @khanacademy/wonder-blocks-modal@3.0.0
+    -   @khanacademy/wonder-blocks-tooltip@1.3.22
+
+## 1.1.21
+
+### Patch Changes
+
+-   Updated dependencies [f3bcc1a9]
+    -   @khanacademy/wonder-blocks-modal@2.3.11
+    -   @khanacademy/wonder-blocks-tooltip@1.3.21
+
 ## 1.1.20
 
 ### Patch Changes