+++ packages/wonder-blocks-tooltip/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @khanacademy/wonder-blocks-tooltip
 
+## 1.3.22
+
+### Patch Changes
+
+-   Updated dependencies [13cdc7fe]
+    -   @khanacademy/wonder-blocks-modal@3.0.0
+
+## 1.3.21
+
+### Patch Changes
+
+-   Updated dependencies [f3bcc1a9]
+    -   @khanacademy/wonder-blocks-modal@2.3.11
+
 ## 1.3.20
 
 ### Patch Changes